# openapi.yaml
openapi: 3.0.3
info:
  title: Multi-Disease Detection System API
  version: "1.0"
  description: >
    Preliminary screening API for 5 diseases (heart, diabetes, breast_cancer,
    kidney, liver). Educational demonstration only — not a certified medical
    device. See /README.md in the repo for the full project write-up.
servers:
  - url: /api

tags:
  - name: auth
  - name: user
  - name: diseases
  - name: predict
  - name: history
  - name: feedback
  - name: admin

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        full_name: { type: string }
        email: { type: string, format: email }
        age: { type: integer, nullable: true }
        gender: { type: string, nullable: true, enum: [male, female, other, null] }
        email_verified: { type: boolean }
        is_admin: { type: boolean }
        created_at: { type: string, format: date-time }
    TestResult:
      type: object
      properties:
        id: { type: string, format: uuid }
        disease: { type: string, enum: [heart, diabetes, breast_cancer, kidney, liver] }
        input_data: { type: object }
        prediction: { type: string }
        probability: { type: number, format: float }
        model_used: { type: string }
        created_at: { type: string, format: date-time }
        explanation:
          type: array
          nullable: true
          description: >
            Top contributing features (SHAP values) for this specific
            prediction, most influential first. `null` if the explanation
            couldn't be computed — the prediction itself is unaffected.
          items:
            type: object
            properties:
              feature: { type: string }
              value: { type: number, description: "The feature's standardized (scaled) value fed to the model" }
              contribution: { type: number, description: "Signed SHAP value — positive pushes toward the positive/disease-present class" }
    Error:
      type: object
      properties:
        error: { type: string }
  parameters:
    DiseasePath:
      name: disease
      in: path
      required: true
      schema: { type: string, enum: [heart, diabetes, breast_cancer, kidney, liver] }
    ResultIdPath:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }

security:
  - bearerAuth: []

paths:
  /health:
    get:
      tags: [diseases]
      summary: Liveness check
      security: []
      responses:
        "200": { description: OK }

  /auth/register:
    post:
      tags: [auth]
      summary: Create an account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [full_name, email, password]
              properties:
                full_name: { type: string }
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
      responses:
        "201":
          description: >
            Account created. Includes a `verification_token` and `dev_note` —
            no email service is configured in this environment, so the
            verification token is returned directly instead of emailed.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  refresh_token: { type: string }
                  user: { $ref: "#/components/schemas/User" }
                  verification_token: { type: string }
                  dev_note: { type: string }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { description: Email already registered }

  /auth/login:
    post:
      tags: [auth]
      summary: Sign in
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  refresh_token: { type: string }
                  user: { $ref: "#/components/schemas/User" }
        "401": { description: Invalid credentials }

  /auth/refresh:
    post:
      tags: [auth]
      summary: Exchange a refresh token for a new access token
      description: Authorization header must carry the *refresh* token, not the access token.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
        "401": { description: Refresh token missing, expired, or revoked }
        "422": { description: Wrong token type (an access token was sent instead of a refresh token) }

  /auth/logout:
    post:
      tags: [auth]
      summary: Revoke a refresh token
      description: >
        Authorization header must carry the refresh token. The paired access
        token remains valid until it naturally expires (it isn't individually
        revocable) — access tokens are short-lived (1h) specifically so this
        window stays small.
      responses:
        "200": { description: Logged out }
        "422": { description: Wrong token type }

  /auth/me:
    get:
      tags: [auth]
      summary: Current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties: { user: { $ref: "#/components/schemas/User" } }
        "401": { description: Missing/invalid access token }

  /auth/verify-email:
    post:
      tags: [auth]
      summary: Confirm the verification token issued at registration
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: { token: { type: string } }
      responses:
        "200": { description: Verified (or already-verified, idempotent) }
        "400": { description: Invalid token }

  /auth/forgot-password:
    post:
      tags: [auth]
      summary: Request a password reset token
      security: []
      description: >
        Always returns the same generic message whether or not the email
        exists (enumeration-safe). If it does exist, a `reset_token` and
        `dev_note` are included — no email service configured, see register.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: { email: { type: string } }
      responses:
        "200": { description: OK (generic message, optionally with a dev-mode token) }

  /auth/reset-password:
    post:
      tags: [auth]
      summary: Reset password using a forgot-password token
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token: { type: string }
                password: { type: string, minLength: 8 }
      responses:
        "200": { description: Password updated }
        "400": { description: Invalid/expired token, or password too short }

  /user/profile:
    get:
      tags: [user]
      summary: Get your profile (same shape as /auth/me)
      responses:
        "200": { description: OK }
    put:
      tags: [user]
      summary: Update full_name / age / gender
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                full_name: { type: string }
                age: { type: integer, nullable: true, minimum: 1, maximum: 120 }
                gender: { type: string, nullable: true, enum: [male, female, other, null] }
      responses:
        "200": { description: Updated }
        "400": { $ref: "#/components/responses/BadRequest" }

  /diseases:
    get:
      tags: [diseases]
      summary: Feature schema + summary metrics for all 5 models
      security: []
      responses:
        "200": { description: OK }

  /diseases/info:
    get:
      tags: [diseases]
      summary: Educational content (overview/symptoms/risk factors/prevention) per disease
      security: []
      responses:
        "200": { description: OK }

  /predict/{disease}:
    post:
      tags: [predict]
      summary: Run a screening from JSON feature values
      parameters:
        - $ref: "#/components/parameters/DiseasePath"
      requestBody:
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        "201":
          description: Prediction saved to history
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TestResult" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { description: Unknown disease key }

  /predict/{disease}/upload:
    post:
      tags: [predict]
      summary: Run a screening from a single-row CSV
      parameters:
        - $ref: "#/components/parameters/DiseasePath"
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties: { file: { type: string, format: binary } }
      responses:
        "201": { description: Prediction saved to history }
        "400": { description: Not a CSV, unparseable, empty, or no recognized columns }
        "404": { description: Unknown disease key }

  /predict/{disease}/upload-pdf:
    post:
      tags: [predict]
      summary: Run a screening from a lab-report PDF (best-effort field extraction)
      description: >
        Regex-based extraction against known feature names — not a general
        document-understanding model. Scanned/image-only PDFs (no text layer)
        aren't supported. Response includes `fields_found` / `fields_missing`
        so the caller knows what was actually detected vs. imputed.
      parameters:
        - $ref: "#/components/parameters/DiseasePath"
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties: { file: { type: string, format: binary } }
      responses:
        "201": { description: Prediction saved to history }
        "400": { description: Not a PDF, unreadable, no text layer, or no recognized fields }
        "404": { description: Unknown disease key }

  /history:
    get:
      tags: [history]
      summary: List your past results
      parameters:
        - name: disease
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TestResult" }

  /history/{id}:
    delete:
      tags: [history]
      summary: Delete one of your results
      parameters:
        - $ref: "#/components/parameters/ResultIdPath"
      responses:
        "200": { description: Deleted }
        "404": { description: Not found (or not yours) }

  /history/{id}/export:
    get:
      tags: [history]
      summary: Download a result as a PDF report
      parameters:
        - $ref: "#/components/parameters/ResultIdPath"
      responses:
        "200":
          description: PDF file
          content:
            application/pdf: { schema: { type: string, format: binary } }
        "404": { description: Not found (or not yours) }

  /feedback:
    post:
      tags: [feedback]
      summary: Submit feedback
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                subject: { type: string, maxLength: 200 }
                message: { type: string, maxLength: 4000 }
      responses:
        "201": { description: Recorded }
        "400": { description: Message missing or too long }

  /admin/users:
    get:
      tags: [admin]
      summary: List all users with their screening counts (admin only)
      responses:
        "200": { description: OK }
        "403": { description: Not an admin }

  /admin/feedback:
    get:
      tags: [admin]
      summary: List all feedback submissions with the submitter's identity (admin only)
      responses:
        "200": { description: OK }
        "403": { description: Not an admin }

  /admin/metrics:
    get:
      tags: [admin]
      summary: Full per-model metrics including confusion matrices (admin only)
      responses:
        "200": { description: OK }
        "403": { description: Not an admin }